requirements.txt
numpy==2.0.0
pandas==2.2.2
paho-mqtt==2.1.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
matplotlib==3.9.2
